Fixed crash caused by parameterised constructor for FXView
Moved the constructor for FXView to a public method that just assigns
the model and controller attributes and then registers observers for
BPM and Beat. It appears that an FX Application class cannot have a
constructor with parameters.

Changed onCloseRequest to system.exit otherwise the BeatBar thread will
keep the application alive.